Allow coverage session to be run stand-alone (#315)
* Allow passing arguments to coverage session

* Allow coverage session to be run stand-alone

Do not combine coverage data when the coverage session is invoked with
arguments, or when there are no data files to be combined.

Combining coverage data is only possible when there are unprocessed
coverage files, such as when triggered by the tests session. This makes
it impossible to use the coverage session on its own. Use cases for that
include the following:

  - Creating an XML coverage report for Codecov during CI
  - Interacting with `coverage` directly during development

Note that this change means that the coverage session is not strictly
idempotent:

  - If there are unprocessed data files, the session combines the files,
    erasing the existing data (and removing the files).

  - If there are no unprocessed data files, the session simply displays
    the existing data.

This behavior appears acceptable because unprocessed data files are
normally only visible during a Nox run, so any coverage session run on
explicit user request _will_ be idempotent.

One case where this becomes relevant is after an aborted Nox run: If
invoked with arguments, the coverage session will show coverage data
from a previous run, or fail if there wasn't a previous run. If invoked
without arguments, the session will show the partial data from the
aborted run.

Rejected ideas:

  Append the coverage data to the combined report in the tests session,
  using `coverage combine --append`. This means coverage data needs to
  be erased explicitly. If you don't remember to erase it, you won't
  notice coverage drops during local testing.

* Document that the coverage session accepts options