Type-check noxfile.py (#342)
* Be strict about types in noxfile

* Type-check noxfile.py against the Python installation running Nox

Use the --python-executable option to mypy to point the type checker to
the Python installation running Nox, rather than the virtual environment
of the Nox session, when type-checking noxfile.py itself. This allows
mypy to find the type information for Nox, without having to install Nox
into the session. Nox distributes type information inline and declares
this fact using the py.typed file specified in PEP 561.

Note that this check is somewhat indeterministic in local testing. The
Nox version is only pinned in CI. For users with a pre-2020.5.24 version
of Nox, the session will fail due to missing type hints. The session may
also conceivably fail in the future if users install a newer version of
Nox with breaking changes. (In the latter case, the failure serves as a
useful reminder that noxfile.py needs to be adapted to the API changes.)

* Eliminate obsolete global variable `locations`