Chain exception correctly (#1015)
According to https://docs.python.org/3/tutorial/errors.html#exception-chaining
And to avoid the error B904 when you will bump flake8-bugbear to a version > 21.9.1, we have to use `raise ... from err` or `raise ... from None`.

As the SysExit() Exception is used I think there is no need to display the exceptions chain. So I propose to use `from None`.